[tests] Move beyond-64k compile tests to integration_test.py
These exercise the top-level compile* functions, not the BaseCompiler
class, so they belong with the other integration tests rather than in a separate
module.
